@@ -1,5 +1,3 @@
 module github.com/charmbracelet/x/exp/maps
 
-go 1.20
-
-require golang.org/x/exp v0.0.0-20231219180239-dc181d75b848
+go 1.21
