@@ -1,9 +1,8 @@
 package maps
 
 import (
+	"slices"
 	"testing"
-
-	"golang.org/x/exp/slices"
 )
 
 func TestSortedKeys(t *testing.T) {
@@ -16,5 +15,18 @@ func TestSortedKeys(t *testing.T) {
 	keys := SortedKeys(m)
 	if slices.Compare(keys, []string{"aaaaa", "bar", "foo"}) != 0 {
 		t.Fatalf("unexpected keys order: %v", keys)
+	}
+}
+
+func TestKeys(t *testing.T) {
+	m := map[string]int{
+		"foo":   1,
+		"bar":   10,
+		"aaaaa": 11,
+	}
+
+	keys := Keys(m)
+	if slices.Compare(keys, []string{"foo", "bar", "aaaaa"}) != 0 {
+		t.Fatalf("unexpected keys: %v", keys)
 	}
 }
